Struct geo_crossings::sweep::Sweep

pub struct Sweep<C: Crossable> where
    C: Crossable + Clone,  {
    segments: Box<Slab<Segment<C>>>,
    events: BinaryHeap<Event<C::Scalar>>,
    active_segments: BTreeSet<ActiveSegment<C>>,
}

Sweep algorithm for detecting all crossings.

This is an internal data-structure that implements the Bentley-Ottman sweep. Maintains a heap of heap of end-points, and the currently active segments. End-users should use one of the iterator interfaces built around this sweep.

Fields
segments: Box<Slab<Segment<C>>>events: BinaryHeap<Event<C::Scalar>>active_segments: BTreeSet<ActiveSegment<C>>
Implementations

impl<C: Crossable + Clone> Sweep<C>

pub fn new<I: IntoIterator<Item = C>>(iter: I) -> Self

fn create_segment(
&mut self,
crossable: C,
geom: Option<LineOrPoint<C::Scalar>>,
parent: Option<usize>
) -> usize

Create a segment, and add its events into the heap.

Arguments

crossable - the user input associated with this segment

geom - the geometry (if None, takes the default geom from crossable)

parent - the chain of overlapping segments to copy from (None if no overlap)

fn adjust_for_intersection(
&mut self,
key: usize,
adj_intersection: LineOrPoint<C::Scalar>
) -> SplitSegments<C::Scalar>

Adjust the segment at key for a detected intersection (adj_intersection).

This is a wrapper around Segment::adjust_for_intersection but also pushes right-end event if the segment geometry changed. Also handles overlaps by adjusting the geom on all overlapping segments.

fn adjust_one_segment(
&mut self,
key: usize,
adj_intersection: LineOrPoint<C::Scalar>
) -> Option<usize>

Adjust the segment at key for a detected intersection (adj_intersection). Splits the segment, and adds the extra segments (copying overlaps appropriately). Returns the key of the overlapping segment, if any.

fn segment_for_event(&self, event: &Event<C::Scalar>) -> Option<&Segment<C>>

Check event is valid, and return the segment associated with it.

If a segment was adjusted, we may have spurious event for the right end point (LineRight) which is no longer valid. Returns None in this case.

fn chain_overlap(&mut self, src_key: usize, tgt_key: usize)

Chain a new segment at tgt_key detected as overlapping with an existing segment at src_key.

fn handle_event<F: FnMut(Crossing<C>)>(
&mut self,
event: Event<C::Scalar>,
cb: &mut F
) -> bool

Handle one event.

Returns true if the event was not spurious.

pub fn next_event<F: FnMut(Crossing<C>)>(
&mut self,
cb: F
) -> Option<SweepPoint<C::Scalar>>

Process the next event in heap.

Calls the callback unless the event is spurious.

pub fn peek_point(&self) -> Option<SweepPoint<C::Scalar>>

Peek and return the next point in the sweep.