Trait geo_crossings::crossings::Crossable

pub trait Crossable: Sized + Debug {
    type Scalar: GeoFloat;
    fn line(&self) -> Line<Self::Scalar>;
}

Interface for types that can be processed to detect crossings.

This type is implemented by Line, but users may also implement this on custom types to store extra information.

Cloning

Note that for usage with the crossing iterators, the type must also impl. Clone. If the custom type is not cheap to clone, use either a reference to the type, a Rc or an Arc. All these are supported via blanket trait implementations.

Associated Types

type Scalar: GeoFloat

Scalar used the coordinates.

Required methods

fn line(&self) -> Line<Self::Scalar>

The geometry associated with this type. Use a Line with the start and end coordinates to represent a point.